requirements.txt
# Core MCP and async
mcp>=0.9.0
httpx>=0.27.0
asyncio>=3.4.3

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Caching
redis>=5.0.0
aioredis>=2.0.1

# RBAC
casbin>=1.36.0
casbin-async-sqlalchemy-adapter>=1.3.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-httpx>=0.43b0
opentelemetry-instrumentation-logging>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
structlog>=24.1.0

# CLI
typer>=0.9.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
respx>=0.20.0
faker>=22.0.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Utilities
tenacity>=8.2.3
